refactor: move all utils functions in separate files
