+++ optimizeImages/lib/utils/index.js
@@ -1,0 +1,6 @@
+export { default as logCompressed } from './logCompressed.js';
+export { default as logConverted } from './logConverted.js';
+export { default as logError } from './logError.js';
+export { default as logVariant } from './logVariant.js';
+export { default as removeExtension } from './removeExtension.js';
+export { default as notImage } from './notImage.js';
